Refactored the main Cosmos client types to have a base abstract class. This will allow users to mock out and unit test there applications.
