+++ Microsoft.Azure.Cosmos/src/Resource/Database/CosmosDatabaseImpl.cs
@@ -1,0 +1,151 @@
+﻿//------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+//------------------------------------------------------------
+
+namespace Microsoft.Azure.Cosmos
+{
+    using System;
+    using System.Net;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using Microsoft.Azure.Cosmos.Internal;
+
+    internal sealed class CosmosDatabaseImpl : CosmosDatabase
+    {
+        internal CosmosDatabaseImpl(
+            CosmosClient client,
+            string databaseId)
+        {
+            this.Id = databaseId;
+            this.Client = client;
+            this.LinkUri = GetLink(null, Paths.DatabasesPathSegment);
+            this.Containers = new CosmosContainersImpl(this);
+        }
+
+        public override CosmosContainers Containers { get; }
+
+        public override string Id { get; }
+
+        internal override CosmosClient Client { get; }
+
+        internal override Uri LinkUri { get; }
+
+        public override Task<CosmosDatabaseResponse> ReadAsync(
+                    CosmosRequestOptions requestOptions = null,
+                    CancellationToken cancellationToken = default(CancellationToken))
+        {
+            Task<CosmosResponseMessage> response = this.ReadStreamAsync(
+                        requestOptions: requestOptions,
+                        cancellationToken: cancellationToken);
+
+            return this.Client.ResponseFactory.CreateDatabaseResponse(this, response);
+        }
+
+        public override Task<CosmosDatabaseResponse> DeleteAsync(
+                    CosmosRequestOptions requestOptions = null,
+                    CancellationToken cancellationToken = default(CancellationToken))
+        {
+            Task<CosmosResponseMessage> response = this.DeleteStreamAsync(
+                        requestOptions: requestOptions,
+                        cancellationToken: cancellationToken);
+
+            return this.Client.ResponseFactory.CreateDatabaseResponse(this, response);
+        }
+
+        public override async Task<int?> ReadProvisionedThroughputAsync(
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            CosmosOfferResult offerResult = await this.ReadProvisionedThroughputIfExistsAsync(cancellationToken);
+            if (offerResult.StatusCode == HttpStatusCode.OK || offerResult.StatusCode == HttpStatusCode.NotFound)
+            {
+                return offerResult.Throughput;
+            }
+
+            throw offerResult.CosmosException;
+        }
+
+        public override async Task ReplaceProvisionedThroughputAsync(
+            int throughput,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            CosmosOfferResult offerResult = await this.ReplaceProvisionedThroughputIfExistsAsync(throughput, cancellationToken);
+            if (offerResult.StatusCode != HttpStatusCode.OK)
+            {
+                throw offerResult.CosmosException;
+            }
+        }
+
+        public override Task<CosmosResponseMessage> ReadStreamAsync(
+                    CosmosRequestOptions requestOptions = null,
+                    CancellationToken cancellationToken = default(CancellationToken))
+        {
+            return this.ProcessAsync(
+                OperationType.Read,
+                requestOptions,
+                cancellationToken);
+        }
+
+        public override Task<CosmosResponseMessage> DeleteStreamAsync(
+                    CosmosRequestOptions requestOptions = null,
+                    CancellationToken cancellationToken = default(CancellationToken))
+        {
+            return this.ProcessAsync(
+                OperationType.Delete,
+                requestOptions,
+                cancellationToken);
+        }
+
+        internal Task<CosmosOfferResult> ReadProvisionedThroughputIfExistsAsync(
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            return this.GetRID(cancellationToken)
+                .ContinueWith(task => 
+                    this.Client.Offers.ReadProvisionedThroughputIfExistsAsync(
+                        task.Result, 
+                        cancellationToken), 
+                    cancellationToken)
+                .Unwrap();
+        }
+
+        internal Task<CosmosOfferResult> ReplaceProvisionedThroughputIfExistsAsync(
+            int targetThroughput,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            Task<string> rid = this.GetRID(cancellationToken);
+            return rid.ContinueWith(task => 
+                this.Client.Offers.ReplaceThroughputIfExistsAsync(
+                    task.Result, 
+                    targetThroughput, 
+                    cancellationToken), 
+                cancellationToken)
+                .Unwrap();
+        }
+
+        internal Task<string> GetRID(CancellationToken cancellationToken = default(CancellationToken))
+        {
+            return this.ReadAsync(cancellationToken: cancellationToken)
+                .ContinueWith(task =>
+                {
+                    CosmosDatabaseResponse response = task.Result;
+                    return response.Resource.ResourceId;
+                }, cancellationToken);
+        }
+
+        private Task<CosmosResponseMessage> ProcessAsync(
+            OperationType operationType,
+            CosmosRequestOptions requestOptions = null,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            return ExecUtils.ProcessResourceOperationStreamAsync(
+                client: this.Client,
+                resourceUri: this.LinkUri,
+                resourceType: ResourceType.Database,
+                operationType: operationType,
+                requestOptions: requestOptions,
+                partitionKey: null,
+                streamPayload: null,
+                requestEnricher: null,
+                cancellationToken: cancellationToken);
+        }
+    }
+}
