+++ Microsoft.Azure.Cosmos/src/Resource/Item/CosmosItems.cs
@@ -18,30 +18,8 @@ namespace Microsoft.Azure.Cosmos
     /// 1. The object operations where it serializes and deserializes the item on request/response
     /// 2. The stream response which takes a Stream containing a JSON serialized object and returns a response containing a Stream
     /// </summary>
-    public class CosmosItems
+    public abstract class CosmosItems
     {
-        /// <summary>
-        /// Cache the full URI segment without the last resource id.
-        /// This allows only a single con-cat operation instead of building the full URI string each time.
-        /// </summary>
-        private string cachedUriSegmentWithoutId { get; }
-        private CosmosJsonSerializer cosmosJsonSerializer { get; }
-        private CosmosClient client { get; }
-
-        /// <summary>
-        /// Create a <see cref="CosmosItems"/>
-        /// </summary>
-        /// <param name="container">The cosmos container</param>
-        protected internal CosmosItems(CosmosContainer container)
-        {
-            this.client = container.Client;
-            this.container = container;
-            this.cosmosJsonSerializer = this.container.Client.CosmosJsonSerializer;
-            this.cachedUriSegmentWithoutId = this.GetResourceSegmentUriWithoutId();
-        }
-
-        internal readonly CosmosContainer container;
-
         /// <summary>
         /// Creates a Item as an asynchronous operation in the Azure Cosmos service.
         /// </summary>
@@ -97,20 +75,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosResponseMessage> CreateItemStreamAsync(
+        public abstract Task<CosmosResponseMessage> CreateItemStreamAsync(
                     object partitionKey,
                     Stream streamPayload,
                     CosmosItemRequestOptions requestOptions = null,
-                    CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return this.ProcessItemStreamAsync(
-                partitionKey,
-                null,
-                streamPayload,
-                OperationType.Create,
-                requestOptions,
-                cancellationToken);
-        }
+                    CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Creates a item as an asynchronous operation in the Azure Cosmos service.
@@ -161,20 +130,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosItemResponse<T>> CreateItemAsync<T>(
+        public abstract Task<CosmosItemResponse<T>> CreateItemAsync<T>(
             object partitionKey,
             T item,
             CosmosItemRequestOptions requestOptions = null,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            Task<CosmosResponseMessage> response = this.CreateItemStreamAsync(
-                partitionKey: partitionKey,
-                streamPayload: this.cosmosJsonSerializer.ToStream<T>(item),
-                requestOptions: requestOptions,
-                cancellationToken: cancellationToken);
-
-            return this.client.ResponseFactory.CreateItemResponse<T>(response);
-        }
+            CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Reads a item from the Azure Cosmos service as an asynchronous operation.
@@ -224,20 +184,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosResponseMessage> ReadItemStreamAsync(
+        public abstract Task<CosmosResponseMessage> ReadItemStreamAsync(
                     object partitionKey,
                     string id,
                     CosmosItemRequestOptions requestOptions = null,
-                    CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return this.ProcessItemStreamAsync(
-                partitionKey,
-                id,
-                null,
-                OperationType.Read,
-                requestOptions,
-                cancellationToken);
-        }
+                    CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Reads a item from the Azure Cosmos service as an asynchronous operation.
@@ -275,20 +226,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosItemResponse<T>> ReadItemAsync<T>(
+        public abstract Task<CosmosItemResponse<T>> ReadItemAsync<T>(
             object partitionKey,
             string id,
             CosmosItemRequestOptions requestOptions = null,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            Task<CosmosResponseMessage> response = this.ReadItemStreamAsync(
-                partitionKey: partitionKey,
-                id: id,
-                requestOptions: requestOptions,
-                cancellationToken: cancellationToken);
-
-            return this.client.ResponseFactory.CreateItemResponse<T>(response);
-        }
+            CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Upserts an item stream as an asynchronous operation in the Azure Cosmos service.
@@ -337,20 +279,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosResponseMessage> UpsertItemStreamAsync(
+        public abstract Task<CosmosResponseMessage> UpsertItemStreamAsync(
                     object partitionKey,
                     Stream streamPayload,
                     CosmosItemRequestOptions requestOptions = null,
-                    CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return this.ProcessItemStreamAsync(
-                partitionKey,
-                null,
-                streamPayload,
-                OperationType.Upsert,
-                requestOptions,
-                cancellationToken);
-        }
+                    CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Upserts an item as an asynchronous operation in the Azure Cosmos service.
@@ -401,20 +334,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosItemResponse<T>> UpsertItemAsync<T>(
+        public abstract Task<CosmosItemResponse<T>> UpsertItemAsync<T>(
             object partitionKey,
             T item,
             CosmosItemRequestOptions requestOptions = null,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            Task<CosmosResponseMessage> response = this.UpsertItemStreamAsync(
-                partitionKey: partitionKey,
-                streamPayload: this.cosmosJsonSerializer.ToStream<T>(item),
-                requestOptions: requestOptions,
-                cancellationToken: cancellationToken);
-
-            return this.client.ResponseFactory.CreateItemResponse<T>(response);
-        }
+            CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Replaces a item in the Azure Cosmos service as an asynchronous operation.
@@ -461,21 +385,12 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosResponseMessage> ReplaceItemStreamAsync(
+        public abstract Task<CosmosResponseMessage> ReplaceItemStreamAsync(
                     object partitionKey,
                     string id,
                     Stream streamPayload,
                     CosmosItemRequestOptions requestOptions = null,
-                    CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return this.ProcessItemStreamAsync(
-                partitionKey,
-                id,
-                streamPayload,
-                OperationType.Replace,
-                requestOptions,
-                cancellationToken);
-        }
+                    CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Replaces a item in the Azure Cosmos service as an asynchronous operation.
@@ -517,22 +432,12 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosItemResponse<T>> ReplaceItemAsync<T>(
+        public abstract Task<CosmosItemResponse<T>> ReplaceItemAsync<T>(
             object partitionKey,
             string id,
             T item,
             CosmosItemRequestOptions requestOptions = null,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            Task<CosmosResponseMessage> response = this.ReplaceItemStreamAsync(
-               partitionKey: partitionKey,
-               id: id,
-               streamPayload: this.cosmosJsonSerializer.ToStream<T>(item),
-               requestOptions: requestOptions,
-               cancellationToken: cancellationToken);
-
-            return this.client.ResponseFactory.CreateItemResponse<T>(response);
-        }
+            CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Delete a item from the Azure Cosmos service as an asynchronous operation.
@@ -569,20 +474,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosResponseMessage> DeleteItemStreamAsync(
+        public abstract Task<CosmosResponseMessage> DeleteItemStreamAsync(
                     object partitionKey,
                     string id,
                     CosmosItemRequestOptions requestOptions = null,
-                    CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return this.ProcessItemStreamAsync(
-                partitionKey,
-                id,
-                null,
-                OperationType.Delete,
-                requestOptions,
-                cancellationToken);
-        }
+                    CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Delete a item from the Azure Cosmos service as an asynchronous operation.
@@ -614,20 +510,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosItemResponse<T>> DeleteItemAsync<T>(
+        public abstract Task<CosmosItemResponse<T>> DeleteItemAsync<T>(
             object partitionKey,
             string id,
             CosmosItemRequestOptions requestOptions = null,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            Task<CosmosResponseMessage> response = this.DeleteItemStreamAsync(
-               partitionKey: partitionKey,
-               id: id,
-               requestOptions: requestOptions,
-               cancellationToken: cancellationToken);
-
-            return this.client.ResponseFactory.CreateItemResponse<T>(response);
-        }
+            CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Gets an iterator to go through all the items for the container
@@ -654,16 +541,9 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator<T> GetItemIterator<T>(
+        public abstract CosmosResultSetIterator<T> GetItemIterator<T>(
             int? maxItemCount = null,
-            string continuationToken = null)
-        {
-            return new CosmosDefaultResultSetIterator<T>(
-                maxItemCount,
-                continuationToken,
-                null,
-                this.ItemFeedRequestExecutor<T>);
-        }
+            string continuationToken = null);
 
         /// <summary>
         /// Gets an iterator to go through all the items for the container as the original CosmosResponseMessage
@@ -694,13 +574,10 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator GetItemStreamIterator(
+        public abstract CosmosResultSetIterator GetItemStreamIterator(
             int? maxItemCount = null,
             string continuationToken = null,
-            CosmosItemRequestOptions requestOptions = null)
-        {
-            return new CosmosDefaultResultSetStreamIterator(maxItemCount, continuationToken, requestOptions, this.ItemStreamFeedRequestExecutor);
-        }
+            CosmosItemRequestOptions requestOptions = null);
 
         /// <summary>
         ///  This method creates a query for items under a container in an Azure Cosmos database using a SQL statement with parameterized values. It returns a CosmosResultSetStreamIterator.
@@ -743,23 +620,12 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator CreateItemQueryAsStream(
+        public abstract CosmosResultSetIterator CreateItemQueryAsStream(
             CosmosSqlQueryDefinition sqlQueryDefinition,
             object partitionKey,
             int? maxItemCount = null,
             string continuationToken = null,
-            CosmosQueryRequestOptions requestOptions = null)
-        {
-            requestOptions = requestOptions ?? new CosmosQueryRequestOptions();
-            Tuple<object, SqlQuerySpec> cxt = new Tuple<object, SqlQuerySpec>(partitionKey, sqlQueryDefinition.ToSqlQuerySpec());
-
-            return new CosmosDefaultResultSetStreamIterator(
-                maxItemCount,
-                continuationToken,
-                requestOptions,
-                this.FeedOrQueryRequestExecutor,
-                cxt);
-        }
+            CosmosQueryRequestOptions requestOptions = null);
 
         /// <summary>
         ///  This method creates a query for items under a container in an Azure Cosmos database using a SQL statement with parameterized values. It returns a CosmosResultSetStreamIterator.
@@ -801,20 +667,12 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator CreateItemQueryAsStream(
+        public abstract CosmosResultSetIterator CreateItemQueryAsStream(
             string sqlQueryText,
             object partitionKey,
             int? maxItemCount = null,
             string continuationToken = null,
-            CosmosQueryRequestOptions requestOptions = null)
-        {
-            return this.CreateItemQueryAsStream(
-                new CosmosSqlQueryDefinition(sqlQueryText),
-                partitionKey,
-                maxItemCount,
-                continuationToken,
-                requestOptions);
-        }
+            CosmosQueryRequestOptions requestOptions = null);
 
         /// <summary>
         ///  This method creates a query for items under a container in an Azure Cosmos database using a SQL statement with parameterized values. It returns a CosmosResultSetIterator.
@@ -850,29 +708,12 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator<T> CreateItemQuery<T>(
+        public abstract CosmosResultSetIterator<T> CreateItemQuery<T>(
             CosmosSqlQueryDefinition sqlQueryDefinition,
             object partitionKey,
             int? maxItemCount = null,
             string continuationToken = null,
-            CosmosQueryRequestOptions requestOptions = null)
-        {
-            CosmosQueryRequestOptions options = requestOptions ?? new CosmosQueryRequestOptions();
-            if (partitionKey != null)
-            {
-                PartitionKey pk = new PartitionKey(partitionKey);
-                options.PartitionKey = pk;
-            }
-
-            options.EnableCrossPartitionQuery = false;
-
-            return new CosmosDefaultResultSetIterator<T>(
-                maxItemCount,
-                continuationToken,
-                options,
-                this.NextResultSetAsync<T>,
-                sqlQueryDefinition.ToSqlQuerySpec());
-        }
+            CosmosQueryRequestOptions requestOptions = null);
 
         /// <summary>
         ///  This method creates a query for items under a container in an Azure Cosmos database using a SQL statement with parameterized values. It returns a CosmosResultSetIterator.
@@ -907,20 +748,12 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator<T> CreateItemQuery<T>(
+        public abstract CosmosResultSetIterator<T> CreateItemQuery<T>(
             string sqlQueryText,
             object partitionKey,
             int? maxItemCount = null,
             string continuationToken = null,
-            CosmosQueryRequestOptions requestOptions = null)
-        {
-            return this.CreateItemQuery<T>(
-                new CosmosSqlQueryDefinition(sqlQueryText),
-                partitionKey,
-                maxItemCount,
-                continuationToken,
-                requestOptions);
-        }
+            CosmosQueryRequestOptions requestOptions = null);
 
         /// <summary>
         ///  This method creates a query for items under a container in an Azure Cosmos database using a SQL statement with parameterized values. It returns a CosmosResultSetIterator.
@@ -957,25 +790,13 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator<T> CreateItemQuery<T>(
+        public abstract CosmosResultSetIterator<T> CreateItemQuery<T>(
             CosmosSqlQueryDefinition sqlQueryDefinition,
             int maxConcurrency,
             int? maxItemCount = null,
             string continuationToken = null,
             CosmosQueryRequestOptions requestOptions = null,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            CosmosQueryRequestOptions options = requestOptions ?? new CosmosQueryRequestOptions();
-            options.maxConcurrency = maxConcurrency;
-            options.EnableCrossPartitionQuery = true;
-
-            return new CosmosDefaultResultSetIterator<T>(
-                maxItemCount,
-                continuationToken,
-                options,
-                this.NextResultSetAsync<T>,
-                sqlQueryDefinition.ToSqlQuerySpec());
-        }
+            CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         ///  This method creates a query for items under a container in an Azure Cosmos database using a SQL statement with parameterized values. It returns a CosmosResultSetIterator.
@@ -1011,241 +832,12 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator<T> CreateItemQuery<T>(
+        public abstract CosmosResultSetIterator<T> CreateItemQuery<T>(
             string sqlQueryText,
             int maxConcurrency,
             int? maxItemCount = null,
             string continuationToken = null,
             CosmosQueryRequestOptions requestOptions = null,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return this.CreateItemQuery<T>(
-                new CosmosSqlQueryDefinition(sqlQueryText),
-                maxConcurrency,
-                maxItemCount,
-                continuationToken,
-                requestOptions,
-                cancellationToken);
-        }
-
-        internal async Task<CosmosQueryResponse<T>> NextResultSetAsync<T>(
-            int? maxItemCount,
-            string continuationToken,
-            CosmosRequestOptions options,
-            object state,
-            CancellationToken cancellationToken)
-        {
-            CosmosQueryRequestOptions cosmosQueryRequestOptions = options as CosmosQueryRequestOptions ?? new CosmosQueryRequestOptions();
-            FeedOptions feedOptions = cosmosQueryRequestOptions.ToFeedOptions();
-            feedOptions.RequestContinuation = continuationToken;
-            feedOptions.MaxItemCount = maxItemCount;
-
-            IDocumentQuery<T> documentClientResult = this.client.DocumentClient.CreateDocumentQuery<T>(
-                collectionLink: this.container.Link,
-                feedOptions: feedOptions,
-                querySpec: state as SqlQuerySpec).AsDocumentQuery();
-
-            try
-            {
-                FeedResponse<T> feedResponse = await documentClientResult.ExecuteNextAsync<T>(cancellationToken);
-                return CosmosQueryResponse<T>.CreateResponse<T>(feedResponse, feedResponse.ResponseContinuation, documentClientResult.HasMoreResults);
-            }
-            catch (DocumentClientException exception)
-            {
-                throw new CosmosException(
-                    message: exception.Message,
-                    statusCode: exception.StatusCode.HasValue ? exception.StatusCode.Value : HttpStatusCode.InternalServerError,
-                    subStatusCode: (int)exception.GetSubStatus(),
-                    activityId: exception.ActivityId,
-                    requestCharge: exception.RequestCharge);
-            }
-        }
-
-        internal virtual Task<CosmosResponseMessage> ProcessItemStreamAsync(
-            object partitionKey,
-            string itemId,
-            Stream streamPayload,
-            OperationType operationType,
-            CosmosRequestOptions requestOptions,
-            CancellationToken cancellationToken)
-        {
-            CosmosItems.ValidatePartitionKey(partitionKey, requestOptions);
-            Uri resourceUri = this.GetResourceUri(requestOptions, operationType, itemId);
-
-            return ExecUtils.ProcessResourceOperationStreamAsync(
-                this.container.Database.Client,
-                resourceUri,
-                ResourceType.Document,
-                operationType,
-                requestOptions,
-                partitionKey,
-                streamPayload,
-                null,
-                cancellationToken);
-        }
-
-        private Task<CosmosResponseMessage> ItemStreamFeedRequestExecutor(
-            int? maxItemCount,
-            string continuationToken,
-            CosmosRequestOptions options,
-            object state,
-            CancellationToken cancellationToken)
-        {
-            Uri resourceUri = this.container.LinkUri;
-            return ExecUtils.ProcessResourceOperationAsync<CosmosResponseMessage>(
-                client: this.container.Database.Client,
-                resourceUri: resourceUri,
-                resourceType: ResourceType.Document,
-                operationType: OperationType.ReadFeed,
-                requestOptions: options,
-                requestEnricher: request =>
-                {
-                    CosmosQueryRequestOptions.FillContinuationToken(request, continuationToken);
-                    CosmosQueryRequestOptions.FillMaxItemCount(request, maxItemCount);
-                },
-                responseCreator: response => response,
-                partitionKey: null,
-                streamPayload: null,
-                cancellationToken: cancellationToken);
-        }
-
-        private Task<CosmosQueryResponse<T>> ItemFeedRequestExecutor<T>(
-            int? maxItemCount,
-           string continuationToken,
-           CosmosRequestOptions options,
-           object state,
-           CancellationToken cancellationToken)
-        {
-            Uri resourceUri = this.container.LinkUri;
-            return ExecUtils.ProcessResourceOperationAsync<CosmosQueryResponse<T>>(
-                client: this.container.Database.Client,
-                resourceUri: resourceUri,
-                resourceType: ResourceType.Document,
-                operationType: OperationType.ReadFeed,
-                requestOptions: options,
-                requestEnricher: request =>
-                {
-                    CosmosQueryRequestOptions.FillContinuationToken(request, continuationToken);
-                    CosmosQueryRequestOptions.FillMaxItemCount(request, maxItemCount);
-                },
-                responseCreator: response => this.client.ResponseFactory.CreateResultSetQueryResponse<T>(response),
-                partitionKey: null,
-                streamPayload: null,
-                cancellationToken: cancellationToken);
-        }
-
-        private Task<CosmosResponseMessage> FeedOrQueryRequestExecutor(
-            int? maxItemCount,
-            string continuationToken,
-            CosmosRequestOptions options,
-            object state,
-            CancellationToken cancellationToken)
-        {
-            if (state == null)
-            {
-                throw new ArgumentNullException(nameof(state));
-            }
-
-            Uri resourceUri = this.container.LinkUri;
-            Tuple<object, SqlQuerySpec> cxt = (Tuple<object, SqlQuerySpec>)state;
-            object partitionKey = cxt.Item1;
-            SqlQuerySpec querySpec = cxt.Item2;
-            if (partitionKey == null && options?.Properties?.TryGetValue(WFConstants.BackendHeaders.EffectivePartitionKeyString, out object effectivePartitionKey) == false)
-            {
-                throw new NotImplementedException(nameof(partitionKey));
-            }
-
-            if (querySpec == null)
-            {
-                throw new NotImplementedException(nameof(querySpec));
-            }
-
-            OperationType queryOperationType = this.client.Configuration.ConnectionMode == ConnectionMode.Direct ? OperationType.Query : OperationType.SqlQuery;
-            Stream streamPayload = this.cosmosJsonSerializer.ToStream(querySpec);
-            return ExecUtils.ProcessResourceOperationAsync<CosmosResponseMessage>(
-                client: this.container.Database.Client,
-                resourceUri: resourceUri,
-                resourceType: ResourceType.Document,
-                operationType: queryOperationType,
-                requestOptions: options,
-                requestEnricher: request =>
-                {
-                    CosmosQueryRequestOptions.FillContinuationToken(request, continuationToken);
-                    CosmosQueryRequestOptions.FillMaxItemCount(request, maxItemCount);
-                },
-                responseCreator: response => response,
-                partitionKey: partitionKey,
-                streamPayload: streamPayload,
-                cancellationToken: cancellationToken);
-        }
-
-        internal Uri GetResourceUri(CosmosRequestOptions requestOptions, OperationType operationType, string itemId)
-        {
-            if (requestOptions != null && requestOptions.TryGetResourceUri(out Uri resourceUri))
-            {
-                return resourceUri;
-            }
-
-            switch (operationType)
-            {
-                case OperationType.Create:
-                case OperationType.Upsert:
-                    return this.container.LinkUri;
-
-                default:
-                    return this.ContcatCachedUriWithId(itemId);
-            }
-        }
-
-        /// <summary>
-        /// Throw an exception if the partition key is null or empty string
-        /// </summary>
-        internal static void ValidatePartitionKey(object partitionKey, CosmosRequestOptions requestOptions)
-        {
-            if (partitionKey != null)
-            {
-                return;
-            }
-
-            if (requestOptions?.Properties != null
-                && requestOptions.Properties.TryGetValue(
-                    WFConstants.BackendHeaders.EffectivePartitionKeyString, out object effectivePartitionKeyValue)
-                && effectivePartitionKeyValue != null)
-            {
-                return;
-            }
-
-            throw new ArgumentNullException(nameof(partitionKey));
-        }
-
-        /// <summary>
-        /// Gets the full resource segment URI without the last id.
-        /// </summary>
-        /// <returns>Example: /dbs/*/colls/*/{this.pathSegment}/ </returns>
-        private string GetResourceSegmentUriWithoutId()
-        {
-            // StringBuilder is roughly 2x faster than string.Format
-            StringBuilder stringBuilder = new StringBuilder(this.container.Link.Length +
-                                                            Paths.DocumentsPathSegment.Length + 2);
-            stringBuilder.Append(this.container.Link);
-            stringBuilder.Append("/");
-            stringBuilder.Append(Paths.DocumentsPathSegment);
-            stringBuilder.Append("/");
-            return stringBuilder.ToString();
-        }
-
-        /// <summary>
-        /// Gets the full resource URI using the cached resource URI segment 
-        /// </summary>
-        /// <param name="resourceId">The resource id</param>
-        /// <returns>
-        /// A document link in the format of {CachedUriSegmentWithoutId}/{0}/ with {0} being a Uri escaped version of the <paramref name="resourceId"/>
-        /// </returns>
-        /// <remarks>Would be used when creating an <see cref="Attachment"/>, or when replacing or deleting a item in Azure Cosmos DB.</remarks>
-        /// <seealso cref="Uri.EscapeUriString"/>
-        private Uri ContcatCachedUriWithId(string resourceId)
-        {
-            return new Uri(this.cachedUriSegmentWithoutId + Uri.EscapeUriString(resourceId), UriKind.Relative);
-        }
+            CancellationToken cancellationToken = default(CancellationToken));
     }
 }
