+++ Microsoft.Azure.Cosmos/src/Resource/StoredProcedure/CosmosStoredProcedures.cs
@@ -14,21 +14,8 @@ namespace Microsoft.Azure.Cosmos
     ///
     /// <see cref="CosmosStoredProcedure"/> for reading, replacing, or deleting an existing stored procedures.
     /// </summary>
-    public class CosmosStoredProcedures
+    public abstract class CosmosStoredProcedures
     {
-        private readonly CosmosContainer container;
-        private readonly CosmosClient client;
-
-        /// <summary>
-        /// Create a <see cref="CosmosStoredProcedures"/>
-        /// </summary>
-        /// <param name="container">The <see cref="CosmosContainer"/> the stored procedures set is related to.</param>
-        protected internal CosmosStoredProcedures(CosmosContainer container)
-        {
-            this.container = container;
-            this.client = container.Client;
-        }
-
         /// <summary>
         /// Creates a stored procedure as an asynchronous operation in the Azure Cosmos DB service.
         /// </summary>
@@ -82,7 +69,7 @@ namespace Microsoft.Azure.Cosmos
         ///        if (!isAccepted) throw new Error(""The query wasn't accepted by the server. Try again/use continuation token between API and script."");
         ///    }";
         ///    
-        /// CosmosStoredProcedure cosmosStoredProcedure = await this.container.StoredProcedures.CreateStoredProcedureAsync(
+        /// CosmosStoredProcedure cosmosStoredProcedure = await this.container.StoredProcedures.CreateStoredProceducreAsync(
         ///         id: "appendString",
         ///         body: sprocBody);
         /// 
@@ -92,39 +79,11 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual Task<CosmosStoredProcedureResponse> CreateStoredProcedureAsync(
+        public abstract Task<CosmosStoredProcedureResponse> CreateStoredProcedureAsync(
                     string id,
                     string body,
                     CosmosRequestOptions requestOptions = null,
-                    CancellationToken cancellationToken = default(CancellationToken))
-        {
-            if (string.IsNullOrEmpty(id))
-            {
-                throw new ArgumentNullException(nameof(id));
-            }
-
-            if (string.IsNullOrEmpty(body))
-            {
-                throw new ArgumentNullException(nameof(body));
-            }
-
-            CosmosStoredProcedureSettings storedProcedureSettings = new CosmosStoredProcedureSettings();
-            storedProcedureSettings.Id = id;
-            storedProcedureSettings.Body = body;
-
-            Task<CosmosResponseMessage> response = ExecUtils.ProcessResourceOperationStreamAsync(
-                this.container.Database.Client,
-                this.container.LinkUri,
-                ResourceType.StoredProcedure,
-                OperationType.Create,
-                requestOptions,
-                partitionKey: null,
-                streamPayload: storedProcedureSettings.GetResourceStream(),
-                requestEnricher: null,
-                cancellationToken: cancellationToken);
-
-            return this.client.ResponseFactory.CreateStoredProcedureResponse(this[id], response);
-        }
+                    CancellationToken cancellationToken = default(CancellationToken));
 
         /// <summary>
         /// Gets an iterator to go through all the stored procedures for the container
@@ -146,16 +105,9 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosResultSetIterator<CosmosStoredProcedureSettings> GetStoredProcedureIterator(
+        public abstract CosmosResultSetIterator<CosmosStoredProcedureSettings> GetStoredProcedureIterator(
             int? maxItemCount = null,
-            string continuationToken = null)
-        {
-            return new CosmosDefaultResultSetIterator<CosmosStoredProcedureSettings>(
-                maxItemCount, 
-                continuationToken, 
-                null, 
-                this.StoredProcedureFeedRequestExecutor);
-        }
+            string continuationToken = null);
 
         /// <summary>
         /// Returns a reference to a stored procedure object. 
@@ -173,35 +125,6 @@ namespace Microsoft.Azure.Cosmos
         /// ]]>
         /// </code>
         /// </example>
-        public virtual CosmosStoredProcedure this[string id]
-        {
-            get
-            {
-                return new CosmosStoredProcedure(this.container, id);
-            }
-        }
-
-        private Task<CosmosQueryResponse<CosmosStoredProcedureSettings>> StoredProcedureFeedRequestExecutor(
-            int? maxItemCount,
-            string continuationToken,
-            CosmosRequestOptions options,
-            object state,
-            CancellationToken cancellationToken)
-        {
-            Uri resourceUri = this.container.LinkUri;
-            return ExecUtils.ProcessResourceOperationAsync<CosmosQueryResponse<CosmosStoredProcedureSettings>>(
-                this.container.Database.Client,
-                resourceUri,
-                ResourceType.StoredProcedure,
-                OperationType.ReadFeed,
-                options,
-                request =>
-                {
-                    CosmosQueryRequestOptions.FillContinuationToken(request, continuationToken);
-                    CosmosQueryRequestOptions.FillMaxItemCount(request, maxItemCount);
-                },
-                response => this.client.ResponseFactory.CreateResultSetQueryResponse<CosmosStoredProcedureSettings>(response),
-                cancellationToken);
-        }
+        public abstract CosmosStoredProcedure this[string id] { get; }
     }
 }
