+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/CosmosClientResourceUnitTests.cs
@@ -27,20 +27,20 @@ namespace Microsoft.Azure.Cosmos.Core.Tests
             string udfId = "udf9003";
 
             CosmosClient mockClient = MockDocumentClient.CreateMockCosmosClient();
-            CosmosDatabase db = new CosmosDatabase(mockClient, databaseId);
-            Assert.AreEqual(db.Link, "/dbs/" + databaseId);
+            CosmosDatabaseImpl db = new CosmosDatabaseImpl(mockClient, databaseId);
+            Assert.AreEqual(db.LinkUri.OriginalString, "/dbs/" + databaseId);
 
-            CosmosContainer container = new CosmosContainer(db, crId);
-            Assert.AreEqual(container.Link, "/dbs/" + databaseId + "/colls/" + crId);
+            CosmosContainerImpl container = new CosmosContainerImpl(db, crId);
+            Assert.AreEqual(container.LinkUri.OriginalString, "/dbs/" + databaseId + "/colls/" + crId);
 
-            CosmosStoredProcedure sp = new CosmosStoredProcedure(container, spId);
-            Assert.AreEqual(sp.Link, "/dbs/" + databaseId + "/colls/" + crId + "/sprocs/" + spId);
+            CosmosStoredProcedureImpl sp = new CosmosStoredProcedureImpl(container, spId);
+            Assert.AreEqual(sp.LinkUri.OriginalString, "/dbs/" + databaseId + "/colls/" + crId + "/sprocs/" + spId);
 
             CosmosTrigger tr = new CosmosTrigger(container, trId);
-            Assert.AreEqual(tr.Link, "/dbs/" + databaseId + "/colls/" + crId + "/triggers/" + trId);
+            Assert.AreEqual(tr.LinkUri.OriginalString, "/dbs/" + databaseId + "/colls/" + crId + "/triggers/" + trId);
 
             CosmosUserDefinedFunction udf = new CosmosUserDefinedFunction(container, udfId);
-            Assert.AreEqual(udf.Link, "/dbs/" + databaseId + "/colls/" + crId + "/udfs/" + udfId);
+            Assert.AreEqual(udf.LinkUri.OriginalString, "/dbs/" + databaseId + "/colls/" + crId + "/udfs/" + udfId);
         }
 
         [TestMethod]
